@@ -9,7 +9,7 @@ import {MatIconModule} from "@angular/material/icon";
 import {MatToolbarModule} from "@angular/material/toolbar";
 import {AboutComponent} from './components/about/about.component';
 import {CaseExplorerComponent} from './components/case-explorer/case-explorer.component';
-import {HttpClientModule} from "@angular/common/http";
+import {HTTP_INTERCEPTORS, HttpClientModule} from "@angular/common/http";
 import {MatSortModule} from "@angular/material/sort";
 import {RegistryViewerComponent} from './components/registry-viewer/registry-viewer.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
@@ -48,6 +48,10 @@ import { HeaderComponent } from './components/header/header.component';
 import {MatDatepickerModule} from "@angular/material/datepicker";
 import {MatTooltipModule} from "@angular/material/tooltip";
 import {APP_DATE_FORMATS, AppDateAdapter} from "./provider/format-datepicker";
+import {AuthHttpInterceptor, AuthModule} from '@auth0/auth0-angular';
+import { Auth0LoginComponent } from './auth0-login/auth0-login.component';
+import {MatMenuModule} from "@angular/material/menu";
+import {MatDividerModule} from "@angular/material/divider";
 
 export const configFactory = (configService: ConfigService) => {
   return () => configService.loadConfig();
@@ -73,7 +77,8 @@ export const configFactory = (configService: ConfigService) => {
     SectionComponent,
     ConformationDialogComponent,
     DemoModeComponent,
-    HeaderComponent
+    HeaderComponent,
+    Auth0LoginComponent
   ],
   imports: [
     BrowserModule,
@@ -105,7 +110,31 @@ export const configFactory = (configService: ConfigService) => {
     MatSnackBarModule,
     MatDatepickerModule,
     NgOptimizedImage,
-    MatTooltipModule
+    MatTooltipModule,
+    AuthModule.forRoot({
+        domain: 'grady-temp.us.auth0.com',
+        clientId: '4T568aTy0dj7keOCla7FubFQO7hJ9iiH',
+        authorizationParams: {
+          redirect_uri: window.location.origin,
+          audience: 'http://smartchartsuite.grady/registry-viewer-api/',
+          scope: 'profile email openid read:scd read:syphilis write:metadata write:scd write:syphilis'
+        },
+        httpInterceptor: {
+          allowedList: [ {
+            uri: 'https://smartchartsuite.dev.heat.icl.gtri.org/registry-viewer-api/*',
+            // tokenOptions: {
+            //   authorizationParams: {
+            //     audience: 'http://smartchartsuite.grady/registry-viewer-api/'
+            //   }
+            // }
+          }
+
+          ]
+        }
+      }
+    ),
+    MatMenuModule,
+    MatDividerModule
   ],
   providers: [
     DrawerService,
@@ -117,7 +146,9 @@ export const configFactory = (configService: ConfigService) => {
       useFactory: configFactory,
       deps: [ConfigService],
       multi: true
-    },],
+    },
+    { provide: HTTP_INTERCEPTORS, useClass: AuthHttpInterceptor, multi: true },
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {
