@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import {AboutComponent} from "./components/about/about.component";
-import {LoginComponent} from "./components/login/login.component";
 import {CaseExplorerComponent} from "./components/case-explorer/case-explorer.component";
 import {RegistryViewerComponent} from "./components/registry-viewer/registry-viewer.component";
+import {LoginComponent} from "./components/login/login.component";
 
 const routes: Routes = [
   //We need to add Login guard to this section
