@@ -4,12 +4,14 @@ import {MatPaginator} from '@angular/material/paginator';
 import {MatSort} from "@angular/material/sort";
 import {ActivatedRoute, Router} from "@angular/router";
 import {CaseRecordsService} from "../../service/case-records.service";
-import {CaseRecord} from "../../domani/case.record";
+import {CaseRecord} from "../../domain/case.record";
 import {FormBuilder, FormGroup} from "@angular/forms";
-import {CaseRecordApiResponse} from "../../domani/case.record.api.response";
+import {CaseRecordApiResponse} from "../../domain/case.record.api.response";
 import {DateAdapter, MAT_DATE_FORMATS} from "@angular/material/core";
 import {APP_DATE_FORMATS, AppDateAdapter} from "../../provider/format-datepicker";
 import {UtilsService} from "../../service/utils.service";
+import {AuthService} from "@auth0/auth0-angular";
+import {combineLatest, mergeMap, of, skipWhile} from "rxjs";
 
 @Component({
   selector: 'app-case-explorer',
@@ -36,12 +38,22 @@ export class CaseExplorerComponent implements OnInit {
     private router: Router,
     private caseRecordsService: CaseRecordsService,
     private formBuilder: FormBuilder,
-    private utilService: UtilsService
+    private utilService: UtilsService,
+    public auth: AuthService
   ) { }
 
   getCaseRecords(searchTerms?: string[]): void {
     this.isLoading = true;
-    this.caseRecordsService.searchCases(searchTerms).subscribe({
+    let search$ = this.caseRecordsService.searchCases(searchTerms);
+    let authenticatedSearch$ = combineLatest(
+      [this.auth.user$, search$]).pipe(
+        skipWhile(combinedResults => combinedResults.some(result => result === undefined)),
+        mergeMap(combinedResults => {
+          console.log(combinedResults[0])
+          return of(combinedResults[1]);
+        })
+    )
+    authenticatedSearch$.subscribe({
       next: (response: CaseRecordApiResponse) => {
         this.dataSource = new MatTableDataSource(response.data);
         this.isLoading = false;
