@@ -1,7 +1,7 @@
 import {Component, Inject, OnInit} from '@angular/core';
 import {UntypedFormBuilder, Validators} from "@angular/forms";
 import {MAT_DIALOG_DATA, MatDialog, MatDialogConfig, MatDialogRef} from "@angular/material/dialog";
-import {Question} from "../../../domani/question";
+import {Question} from "../../../domain/question";
 import {CaseRecordsService} from "../../../service/case-records.service";
 import {UtilsService} from "../../../service/utils.service";
 
@@ -70,7 +70,7 @@ export class AddRecordDialogComponent implements OnInit {
     this.submitted = true;
     this.form.markAllAsTouched();
 
-    //TODO we may need to change this to a specific type and add a class to the domani
+    //TODO we may need to change this to a specific type and add a class to the domain
     if(this.form.valid){
       let keyValueData = {
         manualCaseData:
