@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {Config} from "../domani/config";
+import {Config} from "../domain/config";
 import {HttpBackend, HttpClient} from "@angular/common/http";
 import packageInfo from "../../../package.json";
 import {catchError, map, of} from "rxjs";
